Push compilation branching for lifting into compiler.compile_ir
As title.
